+++ include/geometry.h
@@ -158,5 +158,8 @@ obj_add(struct obj o, ref_t where);
 ref_t
 obj_stack_add(struct obj o, ref_t where,
 		unsigned char n);
+dbref
+contents_find(int descr, ref_t player, ref_t where,
+		const char *name);
 #endif
 #endif
