+++ src/Makefile
@@ -15,8 +15,10 @@ src = ChangeLog COPYING ${real-obj-y:%.o=%.c} compress.c
 
 prochelp-obj-y := prochelp.o
 ws-server-obj-y := ws-server.o
-fbmuck-obj-y := crt_malloc.o biome.o geometry.o noise.o map.o view.o geography.o item.o kill.o shop.o \
-        array.o boolexp.o compile.o create.o db.o db_header.o debugger.o \
+fbmuck-obj-y := geometry.o plant.o biome.o noise.o search.o \
+	view.o item.o spell.o drink.o mob.o kill.o geography.o \
+	shop.o
+fbmuck-obj-y += crt_malloc.o array.o boolexp.o compile.o create.o db.o db_header.o debugger.o \
 	disassem.o diskprop.o edit.o events.o game.o hashtab.o help.o inst.o \
 	interp.o log.o look.o match.o mcp.o mcpgui.o mcppkgs.o mfuns2.o \
 	mfuns.o move.o msgparse.o mufevent.o oldcompress.o p_array.o \
