+++ src/kill.c
@@ -7,27 +7,11 @@
 #include "params.h"
 #include "geography.h"
 #include "view.h"
-#include "map.h"
+#include "search.h"
+#include "spell.h"
+#include "mob.h"
 #undef NDEBUG
 #include "debug.h"
-
-#define DMG_BASE(p) DMG_G(GETSTAT(p, STR))
-#define DODGE_BASE(p) DODGE_G(GETSTAT(p, DEX))
-
-#define HP_G(v) 10 * G(v)
-#define HP_MAX(p) HP_G(GETSTAT(p, CON))
-
-#define MP_G(v) HP_G(v)
-#define MP_MAX(p) MP_G(GETSTAT(p, WIZ))
-
-#define SPELL_G(v) G(v)
-#define SPELL_DMG(p, sp) SPELL_G(GETSTAT(p, INT)) + HS(sp)
-#define SPELL_COST(dmg, y, no_bdmg) (no_bdmg ? 0 : dmg) + dmg / (1 << y)
-
-#define BUF_DURATION(ra) 20 * (RARE_MAX - ra) / RARE_MAX
-#define BUF_DMG(sp_dmg, duration) ((long) 2 * sp_dmg) / duration
-#define BUF_TYPE_MASK 0xf
-#define BUF_TYPE(sp) (sp->flags & BUF_TYPE_MASK)
 
 #define MESGPROP_CXP	"_/cxp"
 #define GETCXP(x)	get_property_value(x, MESGPROP_CXP)
@@ -49,13 +33,6 @@
 #define GETSEATN(x)	get_property_value(x, MESGPROP_SEATN)
 #define SETSEATN(x, y)	set_property_value(x, MESGPROP_SEATN, y)
 
-#define LIVING_SIZE (DB_INITIAL_SIZE/2)
-
-#define HUNGER_Y	4
-#define THIRST_Y	2
-#define HUNGER_INC	(1 << (DAYTICK_Y - HUNGER_Y))
-#define THIRST_INC	(1 << (DAYTICK_Y - THIRST_Y))
-
 struct wts phys_wts[] = {
 	{ "punch", "punches" },
 	{ "peck", "pecks at" },
@@ -64,94 +41,6 @@ struct wts phys_wts[] = {
 	{ "strike", "strikes" },
 };
 
-const char *_spellc[] = {
-	"",
-	ANSI_FG_RED,
-	ANSI_FG_BLUE,
-	ANSI_FG_WHITE,
-	ANSI_FG_YELLOW,
-	ANSI_FG_MAGENTA,
-	ANSI_BOLD ANSI_FG_BLACK,
-};
-
-struct wts buf_wts[] = {
-	// HP
-	{ "heal", "heals" },
-
-	// - HP
-	[16] =
-	{ "bleed", "bleeds" },
-	{ "burn", "burns" },
-
-	// 32 MOV
-
-	// - MOV
-	[48] =
-	{ "stunned", "stunned" },
-	{ "", "" },
-	{ "frozen", "frozen" },
-
-	// MDMG
-	[64] =
-	{ "focus on attacking", "focuses on attacking" },
-	{ "focus on fire", "focuses on fire" },
-
-	// - MDMG
-	[80] =
-	{ "weaken", "weakens" },
-
-	// MDEF
-	[96] =
-	{ "focus on defending", "focuses on defending" },
-	{ "are protected by flames", "is protected by flames" },
-
-	// - MDEF
-	[112] =
-	{ "are distracted", "is distracted" },
-
-	// 128 DODGE
-
-	// 144 - DDGE
-};
-
-static struct _spell spells[] = {
-	{{"Heal", "", "" }, ELM_PHYSICAL, 3, 1, 2, AF_HP },
-
-	{{"Focus", "", "" }, ELM_PHYSICAL, 15, 3, 1, AF_MDMG | AF_BUF, },
-	{{"Fire Focus", "", "" }, ELM_FIRE, 15, 3, 1, AF_MDMG | AF_BUF, },
-
-	{{"Cut", "", "" }, ELM_PHYSICAL, 15, 1, 2, AF_NEG, },
-	{{"Fireball", "", "" }, ELM_FIRE, 3, 1, 2, AF_NEG, }, // 1/4 chance of burning
-
-	{{"Weaken", "", "" }, ELM_PHYSICAL, 15, 3, 1, AF_MDMG | AF_BUF | AF_NEG, },
-	{{"Distract", "", "" }, ELM_PHYSICAL, 15, 3, 1, AF_MDEF | AF_BUF | AF_NEG, },
-
-	{{"Freeze", "", "" }, ELM_ICE, 10, 2, 4, AF_MOV | AF_NEG, },
-
-	{{"Lava Shield", "", "" }, ELM_FIRE, 15, 3, 1, AF_MDEF | AF_BUF, },
-	/* {{"Wind Veil", "", "" }, AF_COMBAT_DODGE, }, */
-	/* {{"Stone Skin", "", "" }, AF_COMBAT_DEFENSE, }, */
-};
-
-enum elm_type weakness[] = {
-	ELM_VAMP, // physical
-	ELM_ICE, // fire
-	ELM_FIRE, // ice
-	ELM_DARK, // air
-	ELM_AIR, // earth
-	ELM_VAMP, // spirit
-	ELM_SPIRIT, // vamp
-	ELM_EARTH, // dark
-};
-
-enum living_flags {
-	LF_NONE,
-	LF_SITTING,
-};
-
-static struct living living[LIVING_SIZE],
-		     *living_cur = &living[0];
-
 static inline short
 randd_dmg(short dmg)
 {
@@ -159,14 +48,14 @@ randd_dmg(short dmg)
 	return xx = dmg + ((dmg * xx * xx * xx) >> 9);
 }
 
-static inline short
-dmg_get(enum elm_type dmg_type, short dmg,
-	short def, enum elm_type def_type)
+short
+kill_dmg(enum element dmg_type, short dmg,
+	short def, enum element def_type)
 {
 	if (dmg > 0) {
-		if (dmg_type == weakness[def_type])
+		if (dmg_type == ELEMENT(def_type)->weakness)
 			dmg *= 2;
-		else if (weakness[dmg_type] == def_type)
+		else if (ELEMENT(dmg_type)->weakness == def_type)
 			dmg /= 2;
 
 		if (dmg < def)
@@ -181,11 +70,11 @@ dmg_get(enum elm_type dmg_type, short dmg,
 
 // returns 1 if target dodges
 static inline int
-dodge_get(struct living *att)
+dodge_get(mobi_t *att)
 {
 	int d = RAND_MAX / 4;
-	short ad = EV(att, DODGE),
-	      dd = EV(att->target, DODGE);
+	short ad = MOBI_EV(att, DODGE),
+	      dd = MOBI_EV(att->target, DODGE);
 
 	if (ad > dd)
 		d += 3 * d / (ad - dd);
@@ -204,18 +93,18 @@ xp_get(unsigned x, unsigned y)
 		return r;
 }
 
-static inline int
-dodge(struct living *att, struct wts wts)
+int
+kill_dodge(mobi_t *att, struct wts wts)
 {
-	struct living *tar = att->target;
-	if (!EV(tar, MOV) && dodge_get(att)) {
+	mobi_t *tar = att->target;
+	if (!MOBI_EV(tar, MOV) && dodge_get(att)) {
 		notify_wts_to(tar->who, att->who, "dodge", "dodges", "'s %s", wts.a);
 		return 1;
 	} else
 		return 0;
 }
 
-static inline void
+void
 notify_attack(dbref att, dbref tar, struct wts wts, short val, char const *color, short mval)
 {
 	char buf[BUFSIZ];
@@ -238,173 +127,14 @@ notify_attack(dbref att, dbref tar, struct wts wts, short val, char const *color
 	notify_wts_to(att, tar, wts.a, wts.b, "%s", buf);
 }
 
-struct living *
-living_get(dbref who)
-{
-	int lid = GETLID(who);
-	CBUG(lid < 0);
-	return &living[lid];
-}
-
-static inline char const *
-sp_color(struct _spell *_sp)
-{
-	return BUF_TYPE(_sp) == AF_HP
-		&& !(_sp->flags & AF_NEG)
-		? ANSI_BOLD ANSI_FG_GREEN
-		: _spellc[_sp->elm_type];
-}
-
-static inline struct wts *
-debuf_wts(struct _spell *_sp)
-{
-	register unsigned char mask = _sp->flags;
-	register unsigned idx = (BUF_TYPE(_sp) << 1) + ((mask >> 4) & 1);
-	idx = (idx << 4) + _sp->elm_type;
-	return &buf_wts[idx];
-}
-
-static void
-debuf_notify(dbref who, struct debuf *d, short val)
-{
-	char buf[BUFSIZ];
-	register struct _spell *_sp = d->_sp;
-	char const *color = sp_color(_sp);
-	struct wts *wts = debuf_wts(_sp);
-
-	if (val)
-		snprintf(buf, sizeof(buf), " (%s%d%s)", color, val, ANSI_RESET);
-	else
-		*buf = '\0';
-
-	notify_wts(who, wts->a, wts->b, "%s", buf);
-}
-
-static inline struct effective *
-eff_get(struct living *liv, unsigned char bt)
-{
-	return &liv->e[bt];
-}
-
-static inline int
-debuf_start(dbref who, struct spell *sp, short val)
-{
-	struct _spell *_sp = sp->_sp;
-	struct living *liv;
-	struct debuf *d;
-	int i;
-
-	liv = living_get(who);
-	if (liv->debuf_mask) {
-		i = __builtin_ffs(~liv->debuf_mask);
-		if (!i)
-			return -1;
-		i--;
-	} else
-		i = 0;
-
-	d = &liv->debufs[i];
-	d->_sp = _sp;
-	d->duration = BUF_DURATION(_sp->ra);
-	d->val = BUF_DMG(val, d->duration);
-
-	i = 1 << i;
-	liv->debuf_mask |= i;
-
-	struct effective *e = eff_get(liv, BUF_TYPE(_sp));
-	e->mask |= i;
-	e->value += d->val;
-
-	debuf_notify(who, d, 0);
-
-	return 0;
-}
-
 static inline void
-spell_init(struct spell *sp, dbref player, unsigned intelligence, unsigned idx)
-{
-	struct _spell *_sp = &spells[idx];
-	sp->_sp = _sp;
-	sp->val = SPELL_DMG(player, _sp);
-	sp->cost = SPELL_COST(sp->val, _sp->y, _sp->flags & AF_BUF);
-}
-
-static inline void
-spells_init(struct spell sps[8], dbref player)
-{
-	char const *str = GETCURSPELLS(player);
-	unsigned i = 0;
-	unsigned intelligence = GETSTAT(player, INT);
-	memset(sps, 0, sizeof(struct spell) * 8);
-
-	if (str && *str != '\0')
-		for (; i < 8; i ++) {
-			char *end;
-			spell_init(&sps[i], player, intelligence,
-			      strtoul(str, &end, 0));
-			str = end;
-			if (*str == '\0') {
-				i++;
-				break;
-			}
-			str++;
-		}
-
-	for (; i < 8; i++)
-		spell_init(&sps[i], player, intelligence, 0);
-}
-
-static inline enum elm_type
-elm_type_next(struct living *liv, register unsigned char a)
-{
-	return a ? liv->debufs[__builtin_ffs(a) - 1]._sp->elm_type
-		: ELM_PHYSICAL;
-}
-
-static inline enum elm_type
-mdef_type(struct living *liv)
-{
-	return elm_type_next(liv, EM(liv, MDEF));
-}
-
-static inline unsigned char
-mdmg_type(struct living *liv)
-{
-	return elm_type_next(liv, EM(liv, MDMG));
-}
-
-static inline void
-debuf_end(struct living *liv, unsigned i)
-{
-	struct debuf *d = &liv->debufs[i];
-	struct effective *e = eff_get(liv, BUF_TYPE(d->_sp));
-	i = 1 << i;
-
-	liv->debuf_mask ^= i;
-	e->mask ^= i;
-	e->value -= d->val;
-}
-
-static inline void
-debufs_end(struct living *liv)
-{
-	register unsigned mask, i, aux;
-
-	for (mask = liv->debuf_mask, i = 0;
-	     (aux = __builtin_ffs(mask));
-	     i++, mask >>= aux)
-
-		 debuf_end(liv, i += aux - 1);
-}
-
-static inline void
-living_dead(struct living *liv)
+living_dead(mobi_t *liv)
 {
 	liv->target = NULL;
 	liv->hp = HP_MAX(liv->who);
 	liv->mp = MP_MAX(liv->who);
 	liv->thirst = liv->hunger = 0;
-	debufs_end(liv);
+	debufs_end(liv->who);
 }
 
 static inline dbref
@@ -460,11 +190,11 @@ xp_award(dbref attacker, dbref target)
 	_xp_award(attacker, xp_get(x, y));
 }
 
-static void
-_kill(dbref attacker, dbref target)
+void
+kill_target(dbref attacker, dbref target)
 {
-	struct living *att = attacker > 0 ? living_get(attacker) : NULL;
-	struct living *tar = living_get(target);
+	mobi_t *att = attacker > 0 ? MOBI(attacker) : NULL;
+	mobi_t *tar = MOBI(target);
 
 	notify_wts(target, "die", "dies", "");
 
@@ -508,96 +238,10 @@ _kill(dbref attacker, dbref target)
 		att->target = NULL;
 }
 
-static int
-hp_add(dbref attacker, dbref target, short amt)
-{
-	struct living *tar = living_get(target);
-	short hp = tar->hp;
-	int ret = 0;
-	hp += amt;
-	if (hp <= 0) {
-		hp = 1;
-		ret = 1;
-		_kill(attacker, target);
-	} else {
-		register unsigned short max = HP_MAX(target);
-		if (hp > max)
-			hp = max;
-	}
-	tar->hp = hp;
-	return ret;
-}
-
-static int
-spell_cast(struct living *att, struct living *tar, unsigned slot)
-{
-	struct spell sp = att->spells[slot];
-	struct _spell *_sp = sp._sp;
-
-	unsigned mana = att->mp;
-	char a[BUFSIZ]; // FIXME way too big?
-	char b[BUFSIZ];
-	char c[BUFSIZ];
-	struct wts wts = { a, b };
-	dbref attacker = att->who,
-	      target = att->target->who;
-
-	char const *color = sp_color(_sp);
-
-	if (mana < sp.cost)
-		return -1;
-
-	snprintf(a, sizeof(a), "%s%s"ANSI_RESET, color, _sp->o.name);
-	memcpy(b, a, sizeof(a));
-
-	mana -= sp.cost;
-	att->mp = mana > 0 ? mana : 0;
-
-	short val = dmg_get(_sp->elm_type, sp.val, EV(tar, MDEF), mdef_type(tar));
-
-	if (_sp->flags & AF_NEG) {
-		val = -val;
-		if (dodge(att, wts))
-			return 0;
-	} else
-		target = attacker;
-
-	snprintf(b, sizeof(b), "casts %s on", a);
-	snprintf(c, sizeof(c), "cast %s on", a);
-	wts.a = c;
-
-	notify_attack(attacker, target, wts, 0, color, val);
-
-	if (random() < (RAND_MAX >> _sp->y))
-		debuf_start(target, &sp, val);
-
-	return hp_add(attacker, target, val);
-}
-
-static inline int
-spells_cast(struct living *att, struct living *tar)
-{
-	register unsigned i, d, combo = att->combo;
-	unsigned enough_mp = 1;
-
-	for (i = 0; enough_mp && (d = __builtin_ffs(combo)); combo >>= d) {
-		switch (spell_cast(att, tar, (i += d) - 1)) {
-		case -1: enough_mp = 0;
-			 break;
-		case 1:  return 1;
-		}
-	}
-
-	if (!enough_mp)
-		notify(att->who, "Not enough mana.");
-
-	return 0;
-}
-
 static inline void
-attack(struct living *p, int player_attack)
+attack(mobi_t *p, int player_attack)
 {
-	struct living *att, *tar;
+	mobi_t *att, *tar;
 	dbref attacker;
 	register unsigned char mask;
 
@@ -611,32 +255,32 @@ attack(struct living *p, int player_attack)
 
 	attacker = att->who;
 
-	mask = EM(att, MOV);
+	mask = MOBI_EM(att, MOV);
 
 	if (mask) {
 		register unsigned i = __builtin_ffs(mask) - 1;
 		debuf_notify(attacker, &att->debufs[i], 0);
-	} else if (!spells_cast(att, tar) && !dodge(att, att->wts)) {
-		enum elm_type at = mdmg_type(att);
-		enum elm_type dt = mdef_type(tar);
+	} else if (!spells_cast(attacker, tar->who) && !kill_dodge(att, att->wts)) {
+		enum element at = ELEMENT_NEXT(att->who, MDMG);
+		enum element dt = ELEMENT_NEXT(tar->who, MDEF);
 		dbref target = tar->who;
-		short aval = -dmg_get(ELM_PHYSICAL, EV(att, DMG), EV(tar, DEF) + EV(tar, MDEF), dt);
-		short bval = -dmg_get(at, EV(att, MDMG), EV(tar, MDEF), dt);
-		char const *color = _spellc[at];
+		short aval = -kill_dmg(ELM_PHYSICAL, MOBI_EV(att, DMG), MOBI_EV(tar, DEF) + MOBI_EV(tar, MDEF), dt);
+		short bval = -kill_dmg(at, MOBI_EV(att, MDMG), MOBI_EV(tar, MDEF), dt);
+		char const *color = ELEMENT(at)->color;
 		notify_attack(attacker, target, att->wts, aval, color, bval);
-		hp_add(attacker, target, aval + bval);
+		cspell_heal(attacker, target, aval + bval);
 	}
 }
 
 static inline void
-target_try(struct living *me, dbref target)
+target_try(mobi_t *me, dbref target)
 {
-	struct living *tar;
+	mobi_t *tar;
 
 	if (me->target)
 		return;
 
-	tar = living_get(target);
+	tar = MOBI(target);
 
 	if (tar == me->target)
 		return;
@@ -651,9 +295,9 @@ do_kill(int descr, dbref player, const char *what)
 	dbref target = strcmp(what, "me")
 		? contents_find(descr, player, here, what)
 		: player;
-	struct living *att, *tar;
+	mobi_t *att, *tar;
 
-        att = GETLID(player) ? living_get(player) : living_put(player);
+        att = GETLID(player) ? MOBI(player) : mob_put(player);
 
 	if (FLAGS(here) & HAVEN
 	    || target == NOTHING
@@ -663,7 +307,7 @@ do_kill(int descr, dbref player, const char *what)
 		 && !((FLAGS(target) & KILL_OK)
 		      && (FLAGS(player) & KILL_OK)))
 #endif
-	    || !(tar = living_get(target)))
+	    || !(tar = MOBI(target)))
 	{
 		notify(player, "You can't target that.");
 		return;
@@ -671,7 +315,7 @@ do_kill(int descr, dbref player, const char *what)
 
 	// this happens when we clone living ones
 	if (tar->who != target)
-		tar = living_put(target);
+		tar = mob_put(target);
 
 	do_stand_silent(player);
 	att->target = tar;
@@ -679,10 +323,10 @@ do_kill(int descr, dbref player, const char *what)
 	notify(player, "You form a combat pose.");
 }
 
-static inline void
-do_killing_update(struct living *liv)
+void
+kill_update(mobi_t *liv)
 {
-	struct living *livt = liv->target;
+	mobi_t *livt = liv->target;
 
 	if (!livt
             || livt == liv
@@ -697,31 +341,20 @@ do_killing_update(struct living *liv)
 	attack(liv, 1);
 }
 
-static void
-respawn(dbref who)
-{
-	if (who < 0 || Typeof(who) != TYPE_THING)
-		return;
-	dbref where = THING_HOME(who);
-	moveto(who, where);
-	notify_except_fmt(DBFETCH(where)->contents, who,
-			  "%s appears.", NAME(who));
-}
-
 void
 do_status(dbref player)
 {
-	// TODO optimize EV / EM
-	struct living *liv = living_get(player);
+	// TODO optimize MOBI_EV / MOBI_EM
+	mobi_t *liv = MOBI(player);
 	notify_fmt(player, "hp %u/%u\tmp %u/%u\tstuck 0x%x\n"
 		   "dodge %d\tcombo 0x%x \tdebuf_mask 0x%x\n"
 		   "damage %d\tmdamage %d\tmdmg_mask 0x%x\n"
 		   "defense %d\tmdefense %d\tmdef_mask 0x%x\n"
 		   "hunger %u\tthirst %u",
-		   liv->hp, HP_MAX(player), liv->mp, MP_MAX(player), EM(liv, MOV),
-		   EV(liv, DODGE), liv->combo, liv->debuf_mask,
-		   EV(liv, DMG), EV(liv, MDMG), EM(liv, MDMG),
-		   EV(liv, DEF), EV(liv, MDEF), EM(liv, MDEF),
+		   liv->hp, HP_MAX(player), liv->mp, MP_MAX(player), MOBI_EM(liv, MOV),
+		   MOBI_EV(liv, DODGE), liv->combo, liv->debuf_mask,
+		   MOBI_EV(liv, DMG), MOBI_EV(liv, MDMG), MOBI_EM(liv, MDMG),
+		   MOBI_EV(liv, DEF), MOBI_EV(liv, MDEF), MOBI_EM(liv, MDEF),
 		   liv->hunger, liv->thirst);
 }
 
@@ -730,7 +363,7 @@ do_heal(int descr, dbref player, const char *name)
 {
 	dbref here = getloc(player);
 	dbref target;
-	struct living *tar;
+	mobi_t *tar;
 
 	if (strcmp(name, "me")) {
 		target = contents_find(descr, player, here, name);
@@ -745,197 +378,12 @@ do_heal(int descr, dbref player, const char *name)
                 return;
         }
 
-        tar = living_get(target);
+        tar = MOBI(target);
 	tar->hp = HP_MAX(target);
 	tar->mp = MP_MAX(target);
 	tar->hunger = tar->thirst = 0;
-	debufs_end(tar);
+	debufs_end(target);
 	notify_wts_to(player, target, "heal", "heals", "");
-}
-
-static inline void
-living_init(struct living *liv, dbref who)
-{
-	int i;
-	living_cur = liv;
-	memset(liv, 0, sizeof(struct living));
-	liv->who = who;
-	liv->mob = mob_get(who);
-	liv->wts = phys_wts[liv->mob ? liv->mob->wt : GETWTS(who)];
-	liv->hunger = liv->thirst = 0;
-	liv->flags = GETSAT(who) ? LF_SITTING : 0;
-	liv->combo = GETCOMBO(who);
-	liv->hp = HP_MAX(who);
-	liv->mp = MP_MAX(who);
-
-	EV(liv, DMG) = DMG_BASE(who);
-	EV(liv, DODGE) = DODGE_BASE(who);
-
-	spells_init(liv->spells, who);
-
-	SETLID(who, living_cur - living);
-
-	for (i = 0; i < EQ_MAX; i++) {
-		register dbref eq = GETEQ(who, i);
-		if (eq > 0)
-			equip_calc(liv, eq);
-	}
-}
-
-struct living *
-living_put(dbref who)
-{
-	register const unsigned m = LIVING_SIZE;
-	register struct living *liv;
-	register unsigned i;
-
-	CBUG(who < 0 || Typeof(who) == TYPE_GARBAGE);
-
-	for (liv = living_cur, i = 0;
-	     i < m;
-	     i++, liv++)
-	{
-		if (liv > living + m)
-			liv = living;
-
-		if (liv->who <= 0) {
-			living_init(liv, who);
-			return liv;
-		}
-	}
-
-	return NULL;
-}
-
-void
-do_living_init()
-{
-	register PropPtr p = get_property((dbref) 0, "_/living");
-        memset(living, 0, sizeof(living));
-
-	if (p) {
-		const char *liv_s = PropDataStr(p), *l;
-		for (l = liv_s; l; l = strchr(l, '#'))
-			living_put((dbref) atoi(++l));
-	}
-}
-
-static inline void
-debufs_process(struct living *liv)
-{
-	register unsigned mask, i, aux;
-	short hpi = 0, dmg;
-	struct debuf *d, *hd;
-
-	for (mask = liv->debuf_mask, i = 0;
-	     (aux = __builtin_ffs(mask));
-	     i++, mask >>= aux)
-	{
-		i += aux - 1;
-		d = &liv->debufs[i];
-		d->duration--;
-		if (d->duration == 0)
-			debuf_end(liv, i);
-		// wtf is this special code?
-		else if (BUF_TYPE(d->_sp) == AF_HP) {
-			dmg = dmg_get(d->_sp->elm_type, d->val,
-				      EV(liv, MDEF), mdef_type(liv));
-			hd = d;
-
-			hpi += dmg;
-		}
-	}
-
-	if (hpi) {
-		debuf_notify(liv->who, hd, hpi);
-		hp_add(NOTHING, liv->who, hpi);
-	}
-}
-
-static inline void
-huth_notify(dbref who, unsigned v, unsigned char y, char const *m[4])
-{
-	static unsigned const n[] = {
-		1 << (DAY_Y - 1),
-		1 << (DAY_Y - 2),
-		1 << (DAY_Y - 3)
-	};
-
-	register unsigned aux;
-
-	if (v == n[2])
-		notify(who, m[0]);
-	else if (v == (aux = n[1]))
-		notify(who, m[1]);
-	else if (v == (aux += n[0]))
-		notify(who, m[2]);
-	else if (v == (aux += n[2]))
-		notify(who, m[3]);
-	else if (v > aux) {
-		short val = -(HP_MAX(who) >> 3);
-		hp_add(NOTHING, who, val);
-	}
-}
-
-static inline void
-living_update(struct living *n)
-{
-	static char const *thirst_msg[] = {
-		"You are thirsty.",
-		"You are very thirsty.",
-		"you are dehydrated.",
-		"You are dying of thirst."
-	};
-
-	static char const *hunger_msg[] = {
-		"You are hungry.",
-		"You are very hungry.",
-		"You are starving.",
-		"You are starving to death."
-	};
-
-	dbref who = n->who;
-
-	if (n->respawn_in > 0) {
-		if (!--n->respawn_in)
-			respawn(who);
-
-		return;
-	} else {
-		int y = 9 - 2 * (n->flags | LF_SITTING);
-
-		int max = HP_MAX(who);
-		int cur = n->hp + (max >> y);
-		n->hp = cur > max ? max : cur;
-
-		max = MP_MAX(who);
-		cur = n->mp + (max >> y);
-		n->mp = cur > max ? max : cur;
-
-		huth_notify(n->who, n->thirst += THIRST_INC,
-			    THIRST_Y, thirst_msg);
-
-		huth_notify(n->who, n->hunger += HUNGER_INC,
-			    HUNGER_Y, hunger_msg);
-	}
-
-	debufs_process(n);
-	do_killing_update(n);
-}
-
-void
-livings_update()
-{
-	register struct living *n, *e;
-
-	for (n = &living[0], e = n + LIVING_SIZE;
-	     n < e;
-	     n++)
-
-		if (n->who > 0)
-			living_update(n);
-		else if (n < living_cur)
-			living_cur = n;
 }
 
 void
@@ -946,13 +394,13 @@ do_advitam(int descr, dbref player, const char *name)
 
 	if (!(FLAGS(player) & WIZARD)
 	    || target == NOTHING
-	    || living_get(target)
+	    || MOBI(target)
 	    || OWNER(target) != player) {
 		notify(player, "You can't do that.");
 		return;
 	}
 
-	living_put(target);
+	mob_put(target);
 	notify_fmt(player, "You infuse %s with life.", NAME(target));
 }
 
@@ -970,38 +418,9 @@ do_givexp(int descr, dbref player, const char *name, const char *amount)
 }
 
 void
-do_living_save()
-{
-	static const unsigned m = LIVING_SIZE;
-	register struct living *n = &living[0];
-	char buf[BUFFER_LEN], *b = buf;
-	size_t len = 0;
-	unsigned i;
-	PData mydat;
-
-	// TODO? use mask?
-	for (i = 0; i < m; i++, n++) {
-		if (n->who > 0 && n->who < db_top && (Typeof(n->who) == TYPE_PLAYER
-				   || Typeof(n->who) == TYPE_THING)
-                    && GETLID(n->who) >= 0)
-		{
-			register size_t d;
-			d = snprintf(b, sizeof(buf) - len, "#%d ", n->who);
-			len += d;
-                        CBUG(len > sizeof(buf));
-			b += d;
-		}
-        }
-
-	mydat.flags = PROP_STRTYP;
-	mydat.data.str = buf;
-	set_property((dbref) 0, "_/living", &mydat);
-}
-
-void
 do_train(dbref player, const char *attrib, const char *amount_s)
 {
-	struct living *liv = living_get(player);
+	mobi_t *liv = MOBI(player);
 	int attr;
 
 	switch (attrib[0]) {
@@ -1028,10 +447,10 @@ do_train(dbref player, const char *attrib, const char *amount_s)
 
 	switch (attr) {
 	case STR:
-		EV(liv, DMG) += DMG_G(c + amount) - DMG_G(c);
+		MOBI_EV(liv, DMG) += DMG_G(c + amount) - DMG_G(c);
 		break;
 	case DEX:
-		EV(liv, DODGE) += DODGE_G(c + amount) - DODGE_G(c);
+		MOBI_EV(liv, DODGE) += DODGE_G(c + amount) - DODGE_G(c);
 		break;
 	}
 
@@ -1046,18 +465,18 @@ kill_v(int descr, dbref player, char const *opcs)
 	if (isdigit(*opcs)) {
 		unsigned combo = strtol(opcs, &end, 0);
 		SETCOMBO(player, combo);
-		living_get(player)->combo = combo;
+		MOBI(player)->combo = combo;
 		notify_fmt(player, "Set combo to 0x%x.", combo);
 		return end - opcs;
 	} else if (*opcs == 'c' && isdigit(opcs[1])) {
-		struct living *p = living_get(player);
+		mobi_t *p = MOBI(player);
 		unsigned slot = strtol(opcs + 1, &end, 0);
 		CBUG(!p);
 
 		if (!p->target)
 			p->target = p;
 
-		spell_cast(p, p->target, slot);
+		spell_cast(p->who, p->target && p->target->who, slot);
 		return end - opcs;
 	} else
 		return 0;
@@ -1092,7 +511,7 @@ do_sit(int descr, dbref player, const char *name)
 
 	SETSEATN(seat, cur + 1);
 	SETSAT(player, seat);
-	living_get(player)->flags |= LF_SITTING;
+	MOBI(player)->flags |= MF_SITTING;
 
 	notify_wts(player, "sit", "sits", " on %s", NAME(seat));
 }
@@ -1109,7 +528,7 @@ do_stand_silent(dbref player)
 		SETSEATN(chair, GETSEATN(chair) - 1);
 
 	USETSAT(player);
-	living_get(player)->flags ^= LF_SITTING;
+	MOBI(player)->flags ^= MF_SITTING;
 	notify_wts(player, "stand", "stands", " up");
 	return 0;
 }
