+++ src/search.c
@@ -1,4 +1,4 @@
-#include "map.h"
+#include "search.h"
 
 #include <db4/db.h>
